Respect global DOCS option
Rename KRB5_PDF and KRB5_HTML options to better reflect subordinate
relationship to DOCS option, and warn when they are no-ops.

Adopt _PORTDOCS helpers and -DOCS-on target to simplify installation
and PLIST generation.

Bump PORTREVISION because PDF and HTML docs are now each installed in
their own subdirectory, and because HTML _sources are no longer
excluded so that search will work.